NVDA compliance, new features
Started using the NVDA addon template to help with addon compliance issues with translations, improve workflow, and ease package geneeration. New features include moving to a global plugin, and the following keyboard gestures:
* alt+windows+k - Enable/disable addon.
While Developer toolkit is enabled:
* Left arrow - Move to previous sibling.
* Up arrow - Move to parent.
* Right arrow - Move to next sibling.
* Down arrow - Move to first child.
* a - List an HTML element's attributes. Press twice quickly to copy to clipboard.
* b - Speak the location of the object's bottom edge. Press twice quickly to copy to clipboard.
* c - Speak the number of children in the focused object. Press twice quickly to copy to clipboard.
* h - Speak the focused object's height. Press twice quickly to copy to clipboard.
* s - Speak the number of siblings to the focused object. Press twice quickly to copy to clipboard.
* w - Speak the focused object's width. Press twice quickly to copy to clipboard.

Also in the process of converting all strings/text to unicode format.
In the process of modularizing code by identifying reusable code and putting it in a shared library.