NavigateTo refactor
Refactor MoveToTopParent and moved common code to the shared.NavigateTo function. MoveToTopParent has a special case that overrides the NavigateTo logic because we are possibly returning an empty list in an attempt to find the top-level parent of all objects on the screen. MoveToTopParent might be a refactor candidate again in the future.