Object's bootom/right edges and relative parent's bottom/right edtes
DTK now allows users the ability to obtain the distance between an object's bottom edge and the bottom edge of the relative parent. Likewise, users can obtain the distance between the object's right edge and the relative parent's right edge.
Cleaned up some relative parent code and added a safety net that sets relative parent to the desktop object when the focused object and relative parent are the same, or when the relative parent is not a direct ancestor of the focused object.