Improved startup routine
Enabling Developer toolkit mode will now set focus to the closest ancestor to the object with focus when the focused object cannot gain focus. For instance, placing the caret on some text not wrapped inside of a span element on a web page will force Developer toolkit to set focus on the next available ancestor element. This continues up the ancestor tree until the document root is reached.