Move to relative parent
Added the ability to move to relative parent by pressing the home key. Note that using numpad7 (home on the numpad) and modifiers such as control, alt, shift, windows, and NVDA along with numpad7 will fail to work, even when using add-ons that restore original functionality to the numpad, the user can only simulate a press of the home key. Unfortunately, these add-ons haven't considered any modifiers the user might include in the key press, and NVDA will still send the original keypress to the operating system. This means that add-on developers must override each keypress and any additional modifiers included.